+++ src/components/screens/Dashboard/index.tsx
@@ -1,8 +1,12 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
+import { useSelector, useDispatch } from "react-redux";
+import { getUserProfile } from "../../../features/profile/profileSlice";
+import { getMyConfigurations } from "../../../features/sim/simSlice";
 import { TradingProfileChart } from "./TradingProfileChart";
 import { ExecutionButtons } from "./ExecutionButtons";
 import "./styles.scss";
 import { SimulationsTable } from "./SimulationsTable";
+import * as secrets from "../../../secrets.js";
 
 interface DashboardProps {}
 
@@ -11,7 +15,26 @@ const initialUser = {
 };
 
 const Dashboard: React.FC<DashboardProps> = () => {
+  const dispatch = useDispatch();
+
   const [user, setUser] = useState(initialUser);
+
+  const getUserProfileWrapper = async () => {
+    const accessToken = secrets.PERMANENT_ACCESS_TOKEN;
+    const username = "LmalcolmR";
+    await dispatch(getUserProfile(accessToken, username));
+  };
+
+  const getUserConfigurationsWrapper = async () => {
+    const accessToken = secrets.PERMANENT_ACCESS_TOKEN;
+    await dispatch(getMyConfigurations(accessToken));
+  };
+
+  useEffect(() => {
+    getUserProfileWrapper();
+    getUserConfigurationsWrapper();
+  }, []);
+
   return (
     <div className="bx--grid">
       <div className="bx--row mb__standard">
