+++ src/secrets.js
@@ -1,4 +1,5 @@
 export const AUTH_SERVER_URL = "auth-server/auth";
-export const SIM_SERVER_URL = 'core-server/graphql';
+export const SIM_SERVER_URL = 'http://localhost:3000/graphql';
 export const IMAGE_SERVER_URL = 'image-server/image';
 export const PROFILE_SERVER_URL = 'http://localhost:4500/profile';
+export const PERMANENT_ACCESS_TOKEN = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiTG1hbGNvbG1SIiwiaWF0IjoxNjAzNjk2NzI1fQ.h5nXMrUYJyWaEP0wkVoxzx3HfM9deYcZHuAgTEL1AkQ';
